requirements.txt
pillow
tkinter
hashlib
csv
